Dashboards: Tighten panel ID types in v2 schema to integers
Change `PanelSpec.id`, `LibraryPanelKindSpec.id`, and `AnnotationPanelFilter.ids`
from CUE `number`/`uint32` to `uint16` across v2alpha1, v2beta1, and v2.

Why: kube-openapi maps `uint32` to `format: int64`, and CUE `number` to
`type: number, format: double`. Consumer codegens that follow the int64-as-string
convention (to avoid JS precision loss past 2^53) then serialise panel IDs as
strings, breaking wire compatibility with Grafana's typed Go decoders. Panel IDs
are integers in practice and always have been; `uint16` narrows the advertised
range to match reality and renders cleanly as `format: int32` in OpenAPI.

Fixes grafana/hyperion-planning#588.

+++ apps/dashboard/pkg/apis/dashboard/v2beta1/dashboard_spec_gen.go
@@ -88,14 +88,14 @@ type DashboardAnnotationPanelFilter struct {
 	// Should the specified panels be included or excluded
 	Exclude *bool `json:"exclude,omitempty"`
 	// Panel IDs that should be included or excluded
-	Ids []uint32 `json:"ids"`
+	Ids []uint16 `json:"ids"`
 }
 
 // NewDashboardAnnotationPanelFilter creates a new DashboardAnnotationPanelFilter object.
 func NewDashboardAnnotationPanelFilter() *DashboardAnnotationPanelFilter {
 	return &DashboardAnnotationPanelFilter{
 		Exclude: (func(input bool) *bool { return &input })(false),
-		Ids:     []uint32{},
+		Ids:     []uint16{},
 	}
 }
 
@@ -180,7 +180,7 @@ func (DashboardPanelKind) OpenAPIModelName() string {
 
 // +k8s:openapi-gen=true
 type DashboardPanelSpec struct {
-	Id          float64                 `json:"id"`
+	Id          uint16                  `json:"id"`
 	Title       string                  `json:"title"`
 	Description string                  `json:"description"`
 	Links       []DashboardDataLink     `json:"links"`
@@ -1032,7 +1032,7 @@ func (DashboardLibraryPanelKind) OpenAPIModelName() string {
 // +k8s:openapi-gen=true
 type DashboardLibraryPanelKindSpec struct {
 	// Panel ID for the library panel in the dashboard
-	Id float64 `json:"id"`
+	Id uint16 `json:"id"`
 	// Title for the library panel in the dashboard
 	Title        string                   `json:"title"`
 	LibraryPanel DashboardLibraryPanelRef `json:"libraryPanel"`
